LOG-3436: Refactor logstore into separate package
