+++ internal/k8shandler/status.go
@@ -57,92 +57,6 @@ func (clusterRequest *ClusterLoggingRequest) getKibanaStatus() ([]elasticsearch.
 	return cr.Status, nil
 }
 
-func (clusterRequest *ClusterLoggingRequest) getElasticsearchStatus() ([]logging.ElasticsearchStatus, error) {
-
-	// we can scrape the status provided by the elasticsearch-operator
-	// get list of elasticsearch objects
-	esList := &elasticsearch.ElasticsearchList{
-		TypeMeta: metav1.TypeMeta{
-			Kind:       "Elasticsearch",
-			APIVersion: elasticsearch.GroupVersion.String(),
-		},
-	}
-
-	err := clusterRequest.List(map[string]string{}, esList)
-	status := []logging.ElasticsearchStatus{}
-
-	if err != nil {
-		return status, fmt.Errorf("Unable to get Elasticsearches: %v", err)
-	}
-
-	if len(esList.Items) != 0 {
-		for _, cluster := range esList.Items {
-			nodeConditions := make(map[string]logging.ElasticsearchClusterConditions)
-
-			nodeStatus := logging.ElasticsearchStatus{
-				ClusterName:            cluster.Name,
-				NodeCount:              cluster.Status.Cluster.NumNodes,
-				ClusterHealth:          cluster.Status.ClusterHealth,
-				Cluster:                cluster.Status.Cluster,
-				Pods:                   getPodMap(cluster.Status),
-				ClusterConditions:      logging.ElasticsearchClusterConditions(cluster.Status.Conditions),
-				ShardAllocationEnabled: cluster.Status.ShardAllocationEnabled,
-			}
-
-			for _, node := range cluster.Status.Nodes {
-				nodeName := ""
-
-				if node.DeploymentName != "" {
-					nodeName = node.DeploymentName
-				}
-
-				if node.StatefulSetName != "" {
-					nodeName = node.StatefulSetName
-				}
-
-				if node.Conditions != nil {
-					nodeConditions[nodeName] = logging.ElasticsearchClusterConditions(node.Conditions)
-				} else {
-					nodeConditions[nodeName] = []elasticsearch.ClusterCondition{}
-				}
-			}
-
-			nodeStatus.NodeConditions = nodeConditions
-
-			status = append(status, nodeStatus)
-		}
-	}
-
-	return status, nil
-}
-
-func getPodMap(node elasticsearch.ElasticsearchStatus) map[logging.ElasticsearchRoleType]logging.PodStateMap {
-
-	return map[logging.ElasticsearchRoleType]logging.PodStateMap{
-		logging.ElasticsearchRoleTypeClient: translatePodMap(node.Pods[elasticsearch.ElasticsearchRoleClient]),
-		logging.ElasticsearchRoleTypeData:   translatePodMap(node.Pods[elasticsearch.ElasticsearchRoleData]),
-		logging.ElasticsearchRoleTypeMaster: translatePodMap(node.Pods[elasticsearch.ElasticsearchRoleMaster]),
-	}
-}
-
-func translatePodMap(podStateMap elasticsearch.PodStateMap) logging.PodStateMap {
-
-	return logging.PodStateMap{
-		logging.PodStateTypeReady:    getPodState(podStateMap, elasticsearch.PodStateTypeReady),
-		logging.PodStateTypeNotReady: getPodState(podStateMap, elasticsearch.PodStateTypeNotReady),
-		logging.PodStateTypeFailed:   getPodState(podStateMap, elasticsearch.PodStateTypeFailed),
-	}
-}
-
-func getPodState(podStateMap elasticsearch.PodStateMap, podStateType elasticsearch.PodStateType) []string {
-	if v, ok := podStateMap[podStateType]; ok {
-		sort.Strings(v)
-		return v
-	} else {
-		return []string{}
-	}
-}
-
 func podStateMap(podList []v1.Pod) logging.PodStateMap {
 	stateMap := map[logging.PodStateType][]string{
 		logging.PodStateTypeReady:    {},
