+++ internal/logstore/elasticsearch/indexmanagement/index_management.go
