+++ internal/logstore/elasticsearch/suite_test.go
@@ -1,0 +1,13 @@
+package elasticsearch_test
+
+import (
+	"testing"
+
+	. "github.com/onsi/ginkgo"
+	. "github.com/onsi/gomega"
+)
+
+func TestSuite(t *testing.T) {
+	RegisterFailHandler(Fail)
+	RunSpecs(t, "[internal][logstore][elasticsearch] suite")
+}
